Cherrygram ver. 10.7.0
🇺🇸 Changelog:
• 🆕 Universal QR-Reader,
• 🆕 "Open in external app" button for documents,
— Bug fixes and performance improvements.

🇷🇺 Список изменений:
• 🆕 Универсальный сканер QR-кодов,
• 🆕 Кнопка "Открыть во внешнем приложении" для файлов,
— Исправления багов и прочие улучшения.

(cherry picked from commit 621d04f15c902a49f1188dd7d3f7fa0ab3d70eb1)

+++ TMessagesProj/src/main/java/org/telegram/messenger/MediaController.java
@@ -740,6 +740,7 @@ public class MediaController implements AudioManager.OnAudioFocusChangeListener,
             final Pair<Integer, Integer> orientation = AndroidUtilities.getImageOrientation(filterPath != null ? filterPath : path);
             final Bitmap.CompressFormat compressFormat = Bitmap.CompressFormat.JPEG;
             final Bitmap bitmap = StoryEntry.getScaledBitmap(opts -> BitmapFactory.decodeFile(filterPath != null ? filterPath : path, opts), AndroidUtilities.getPhotoSize(highQuality), AndroidUtilities.getPhotoSize(highQuality), false, true);
+            if (bitmap == null) return;;
             if (imagePath != null) {
                 new File(imagePath).delete(); imagePath = null;
             }
